Split 'block.js' into smaller chunks
