+++ assets/js/blocks/reviews-by-product/components/load-more-button.js
@@ -1,0 +1,18 @@
+/**
+ * External dependencies
+ */
+import { __ } from '@wordpress/i18n';
+
+/**
+ * Internal dependencies
+ */
+export const LoadMoreButton = ( { onClick } ) => (
+	<button
+		className="wc-block-reviews-by-product__load-more"
+		onClick={ onClick }
+	>
+		{ __( 'Load more', 'woo-gutenberg-products-block' ) }
+	</button>
+);
+
+export default LoadMoreButton;
